refactor: modularize realtime server
